requirements.txt
aws-cdk-lib==2.118.0
constructs==10.3.0
dynaconf==3.2.4